Move rand function to module core::math::rand
