+++ src/main.rs
@@ -11,11 +11,11 @@ use crate::core::camera::Camera;
 use crate::core::color::write_color;
 use crate::core::color::Color;
 use crate::core::geometry::{Point3, Ray, Vec3, Vector};
+use crate::core::math::rand::{rand, rand_between};
 use crate::materials::{Dielectric, Lambertian, Material, Metal};
 use crate::scene::{Hit, HitList};
 use crate::sphere::Sphere;
 use crate::util::INFINITY;
-use crate::util::{rand, rand_between};
 
 fn ray_color(ray: Ray, world: &HitList, depth: i32) -> Color {
     // Stop recursion at ray bounce limit
