+++ src/materials/lambertian.rs
@@ -1,7 +1,7 @@
 use crate::core::color::Color;
 use crate::core::geometry::Ray;
+use crate::core::math::rand::rand_unit_vector;
 use crate::scene::BasicHitRecord;
-use crate::util::rand_unit_vector;
 
 use super::material::{Material, ScatterRecord};
